Move tasks into its own package
