+++ src/main/java/deez/tasks/Todo.java
@@ -1,4 +1,4 @@
-package deez;
+package deez.tasks;
 
 /**
  * A class representing a Todo task.
